Add validation to transition type
This adds validations to ensure that:

1. The value of `resource` is either a valid resource reference or a
string that can be converted to a valid resource reference.
2. The resource specified in `resource` exists in the catalog.
3. The value of `attributes` is a hash.
4. All the specified attributes are valid parameters for the specified
resource.
5. The values of `prior_to` are valid resource references or strings
that can be converted to valid resource references.
6. All resources specified in `prior_to` exist in the catalog.

Additionally, it adds munging to resolve the values of `resource` and
`prior_to` into their catalog representations and removes the
now-redundant resource resolution from the provider.